Skip flakey grpc watch test
Summary:
Title. Not sure what's going on here. Ran this for 50 times where it was fine, but occasionally we don't see the event we want to see.

We've since introduced timeouts on the gRPC calls, but there might be something going on that is causing the threads to execute slower than we would like them on BK. Will investigate using tox and the bk image.

Test Plan: unit

Reviewers: johann

Reviewed By: johann

Differential Revision: https://dagster.phacility.com/D6398