[dagit] Consistent blue color for nav, search
Summary: I've been using a brighter `COBALT`-based set of colors for some of the newer UI, but this doesn't match the `BLUE`-based colors used elsewhere. The `BLUE` blues are a little deader than the `COBALT` blues, but for now I think it's better to maintain consistency, and we can revisit colors globally later on.

Test Plan: Toggle on left nav, verify that the blues match. Open search, verify same.

Reviewers: bengotow

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6402