[dagit] Don't block viewing PythonError on Playground preview
Summary:
When an error occurs in generating the run preview pane on the Dagit playground, we have a PythonError but aren't correctly clearing the loading overlay. Fix this by only showing the overlay when the preview is actually loading. I'm not sure if I introduced this bug recently, but hopefully I'm not overlooking anything here.

Also change the error link to be a proper ButtonLink.

Test Plan: Yield a `SerializableErrorInfo` from `GrpcServerRepositoryLocation.get_subset_external_pipeline_result`, open playground and select a solid. Verify that the error is clickable and not blocked by the loading overlay.

Reviewers: dgibson

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D6435