Add wrapper around pendulum that works with both pendulum 2 and pendulum 1
Summary:
I don't love this solution and i'm still figuring a testing solution (separate tox suite for pendulum 1 vs. pendulum 2?), but as long as airflow 1 is on an older version of pendulum and airflow 2 is on a newer version, and we want both to work with dagster, it seems like we need some kind of shim like this on top of pendulum. Most of the changes from pendulum 1 to pendulum 2 are pretty cosmetic, as reflected here.

Does this overall strategy seem resaonable? If so I can clean it up and add the additional test coverage for the old version.

Test Plan: BK (with new test suite)

Reviewers: prha, max, alangenfeld, nate

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D6550